Extract resources repack step
